Move lock tracking to shop container
